+++ src/app/components/LandingPage/VoTSection.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '@/components/Button'
+import Button from '@/components/ui/Button'
 
 interface VoTSectionProps {
   votRef: React.RefObject<HTMLElement | null>
