+++ src/app/components/UploadModal.tsx
@@ -4,10 +4,10 @@ import React, { useState, useCallback, useMemo, useRef } from 'react'
 
 import { cn, SIZE_CLASSES } from '@/lib/utils'
 
-import Modal, { type ModalProps } from './Modal'
-import Button from './Button'
-import ButtonGroup from './ButtonGroup'
-import ProgressCircle from './ProgressCircle'
+import Modal, { type ModalProps } from '@/components/ui/Modal'
+import Button from '@/components/ui/Button'
+import ButtonGroup from '@/components/ui/ButtonGroup'
+import ProgressCircle from '@/components/ui/ProgressCircle'
 import FileUploadTab from './UploadModal/FileUploadTab'
 import UrlImportTab from './UploadModal/UrlImportTab'
 import TranscriptionSettings from './UploadModal/TranscriptionSettings'
