+++ src/app/components/UploadModal/UrlImportTab.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import AlertBanner from '@/components/AlertBanner'
+import AlertBanner from '@/components/shared/AlertBanner'
 
 interface UrlImportTabProps {
   url: string
