+++ src/app/editor/components/EditorToolbar.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import Button from '@/components/Button'
+import Button from '@/components/ui/Button'
 
 export interface EditorToolbarProps {
   activeTab: string
