+++ src/app/page.tsx
@@ -4,15 +4,15 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { useCallback, useEffect, useRef, useState } from 'react'
 
-import Button from '@/components/Button'
-import EditTranscriptionSection from '@/components/LandingPage/EditTranscriptionSection'
-import FastTranscriptionSection from '@/components/LandingPage/FastTranscriptionSection'
-import Footer from '@/components/LandingPage/Footer'
-import HeroSection from '@/components/LandingPage/HeroSection'
-import OpenLibrarySection from '@/components/LandingPage/OpenLibrarySection'
-import SubtitleEditorSection from '@/components/LandingPage/SubtitleEditorSection'
-import VoTSection from '@/components/LandingPage/VoTSection'
-import UploadModal from '@/components/UploadModal'
+import Button from '@/components/ui/Button'
+import EditTranscriptionSection from './components/LandingPage/EditTranscriptionSection'
+import FastTranscriptionSection from './components/LandingPage/FastTranscriptionSection'
+import Footer from './components/LandingPage/Footer'
+import HeroSection from './components/LandingPage/HeroSection'
+import OpenLibrarySection from './components/LandingPage/OpenLibrarySection'
+import SubtitleEditorSection from './components/LandingPage/SubtitleEditorSection'
+import VoTSection from './components/LandingPage/VoTSection'
+import UploadModal from './components/UploadModal'
 import { useUploadModal } from '@/hooks/useUploadModal'
